+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - update to `sdk-rs 0.45.2`
 
+### Fixed
+
+- check number of tokens during deployment
+
 ## [1.0.0]
 
 ### Added